Added options to override physical directories for www, www_android, www_ios, www_wp8 etc. Merged lib/middleware/cordova/* -> lib/middleware/cordova.js and spec/middleware/cordova.spec.js
