+++ lib/middleware/static.js
@@ -11,6 +11,6 @@ var connect = require('connect'),
  * Serves the PhoneGap app from the current working directory.
  */
 
-module.exports = function() {
-    return connect.static(path.join(process.cwd(), 'www'));
+module.exports = function(options) {
+    return connect.static(path.resolve(options.www || 'www'));
 };
